physical-ai-book/backend/rag_api/requirements.txt
# RAG API for Physical AI Textbook
# FastAPI-based chatbot with vector search

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
qdrant-client==1.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
openai==1.7.2
sentence-transformers==2.2.2
numpy==1.26.3
httpx==0.26.0
python-multipart==0.0.6
fastapi-cors==0.0.6
